# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ToolRoute API",
    "description": "The OpenRouter for Tools — unified API gateway for 50+ AI agent tools. One API key, every tool.",
    "version": "1.0.0",
    "contact": {
      "url": "https://toolroute.ai",
      "email": "support@toolroute.ai"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://toolroute.ai",
      "description": "Production"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/v1/tools": {
      "get": {
        "operationId": "listToolsV1",
        "summary": "List available tools with pricing",
        "description": "Returns all tools in the ToolRoute registry with pricing, capabilities, and provider information. No authentication required.",
        "tags": ["Discovery"],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by super_category slug"
          },
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search query to filter tools"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 },
            "description": "Max results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "List of available tools",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tools": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ToolListing" }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/keys": {
      "post": {
        "operationId": "createApiKey",
        "summary": "Create a new API key",
        "description": "Creates a new API key for the authenticated user. Requires session authentication (logged in via dashboard).",
        "tags": ["Keys"],
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Human-readable label for this key",
                    "example": "My Agent Key"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created. The full key is only returned once.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The full API key (only shown once)",
                      "example": "tr_live_abc123def456"
                    },
                    "id": { "type": "string", "format": "uuid" },
                    "name": { "type": "string" },
                    "created_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listApiKeys",
        "summary": "List API keys",
        "description": "Lists all API keys for the authenticated user. Keys are masked (only last 4 characters shown).",
        "tags": ["Keys"],
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "keys": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "key_hint": {
                            "type": "string",
                            "description": "Masked key showing only last 4 chars",
                            "example": "tr_live_...f456"
                          },
                          "created_at": { "type": "string", "format": "date-time" },
                          "last_used_at": { "type": "string", "format": "date-time", "nullable": true },
                          "total_requests": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "revokeApiKey",
        "summary": "Revoke an API key",
        "description": "Permanently revokes an API key. This cannot be undone.",
        "tags": ["Keys"],
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["key_id"],
                "properties": {
                  "key_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the key to revoke"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Key revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revoked": { "type": "boolean" },
                    "key_id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Key not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/usage": {
      "get": {
        "operationId": "getUsageHistory",
        "summary": "Get usage history",
        "description": "Returns usage history for the authenticated API key or session. Includes per-request cost, tool used, latency, and timestamps.",
        "tags": ["Usage"],
        "security": [{ "apiKey": [] }, { "sessionAuth": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 },
            "description": "Max records to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 },
            "description": "Offset for pagination"
          },
          {
            "name": "tool",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by tool identifier (e.g. resend/send_email)"
          },
          {
            "name": "since",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "Only return usage after this timestamp"
          }
        ],
        "responses": {
          "200": {
            "description": "Usage history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usage": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/UsageRecord" }
                    },
                    "total": { "type": "integer" },
                    "total_cost": { "type": "number", "description": "Sum of cost for all matching records" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key in the format tr_live_xxx. Pass as Bearer token in Authorization header."
      },
      "sessionAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Supabase JWT session token from dashboard login. Pass as Bearer token in Authorization header."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Human-readable error message",
                "example": "Insufficient balance. Add credits at https://toolroute.ai/dashboard/billing"
              },
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "enum": [
                  "invalid_request",
                  "authentication_required",
                  "invalid_api_key",
                  "insufficient_balance",
                  "tool_not_found",
                  "rate_limit_exceeded",
                  "provider_error",
                  "internal_error"
                ],
                "example": "insufficient_balance"
              }
            }
          }
        }
      },
      "ToolListing": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "example": "Resend" },
          "slug": { "type": "string", "example": "resend" },
          "description": { "type": "string" },
          "operations": {
            "type": "array",
            "description": "Available operations for this tool",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "example": "resend/send_email" },
                "name": { "type": "string", "example": "Send Email" },
                "description": { "type": "string" },
                "price_per_call": { "type": "number", "example": 0.001 }
              }
            }
          },
          "rating": { "type": "integer", "minimum": 1, "maximum": 10 },
          "super_category": { "type": "string" },
          "sub_category": { "type": "string" },
          "protocols": {
            "type": "array",
            "items": { "type": "string", "enum": ["rest", "mcp", "graphql", "grpc", "sdk"] }
          },
          "cost": {
            "type": "string",
            "enum": ["free", "freemium", "paid", "usage-based", "enterprise"]
          },
          "status": { "type": "string", "enum": ["active", "beta", "deprecated"] }
        }
      },
...
      "UsageRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "tool": { "type": "string", "example": "resend/send_email" },
          "provider": { "type": "string", "example": "resend" },
          "cost": { "type": "number", "example": 0.001 },
          "latency_ms": { "type": "integer", "example": 234 },
          "status": { "type": "string", "enum": ["success", "error"] },
          "error_code": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
}